Add RADONEYE_ROUNDING_OFF debug feature

+++ src/radoneye/model.py
@@ -6,13 +6,13 @@ class RadonEyeStatus(TypedDict):
     serial: str
     model: str
     version: str
-    latest_bq_m3: int
+    latest_bq_m3: float
     latest_pci_l: float
-    day_avg_bq_m3: int
+    day_avg_bq_m3: float
     day_avg_pci_l: float
-    month_avg_bq_m3: int
+    month_avg_bq_m3: float
     month_avg_pci_l: float
-    peak_bq_m3: int
+    peak_bq_m3: float
     peak_pci_l: float
     counts_current: int
     counts_previous: int
@@ -26,7 +26,7 @@ class RadonEyeStatus(TypedDict):
 
 
 class RadonEyeHistory(TypedDict):
-    values_bq_m3: list[int]
+    values_bq_m3: list[float]
     values_pci_l: list[float]
